new-dhanush-AI-main/requirements.txt
# -------------------------------
# Core Flask & Database
# -------------------------------
Flask==3.0.3
Werkzeug==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.9.0
Jinja2==3.1.4

# -------------------------------
# NLP / Resume Parsing
# -------------------------------
# spaCy (3.9.x) not yet supports Python 3.13 — skip or install manually when wheel released
sentence-transformers==3.1.1
pdfminer.six==20240706

PyMuPDF==1.24.10
python-docx==0.8.11
docx2txt==0.8
numpy==2.1.3
pandas==2.2.3

# -------------------------------
# AI / ML Models
# -------------------------------
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
torch==2.4.1
transformers==4.45.2
vaderSentiment==3.3.2
textblob==0.17.1

# -------------------------------
# AI Interview & Audio (Optional)
# -------------------------------
python-multipart==0.0.9
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1

# -------------------------------
# AI Frameworks & LangChain
# -------------------------------
langchain==0.3.4
chromadb==0.5.5

# -------------------------------
# Optional Integrations
# -------------------------------
openai==1.50.2
prophet==1.1.5          # may require C++ build tools on Windows
marian-nmt              # optional Hugging Face translation wrapper
requests==2.32.3
